Crate re_data_ui

Rerun Data Ui

This crate provides ui elements for Rerun component data for the Rerun Viewer.

Modules

annotation_context ğŸ”’

component ğŸ”’

component_path ğŸ”’

component_ui_registry ğŸ”’

data ğŸ”’

entity_path ğŸ”’

image ğŸ”’

instance_path ğŸ”’

item ğŸ”’

item_ui

Basic ui elements & interaction for most re_viewer_context::Item.

log_msg ğŸ”’

pinhole ğŸ”’

transform3d ğŸ”’

Traits

DataUi

Types implementing DataUi can display themselves in an [egui::Ui].

EntityDataUi

Similar to DataUi, but for data that is related to an entity (e.g. a component).

Functions

annotations

create_component_ui_registry

f ğŸ”’

format_cell ğŸ”’

show_zoomed_image_region

meter: iff this is a depth map, how long is one meter?

show_zoomed_image_region_area_outline

tensor_summary_ui_grid_contents