Module re_data_ui::image

Constants

ZOOMED_IMAGE_TEXEL_RADIUS ğŸ”’

Functions

copy_and_save_image_ui ğŸ”’

format_tensor_shape_single_line

largest_size_that_fits_in ğŸ”’

rgb8_histogram_ui ğŸ”’

save_image ğŸ”’

show_image_preview ğŸ”’

Shows preview of an image.

show_zoomed_image_region

meter: iff this is a depth map, how long is one meter?

show_zoomed_image_region_area_outline

show_zoomed_image_region_tooltip ğŸ”’

tensor_pixel_value_ui ğŸ”’

tensor_summary_ui

tensor_summary_ui_grid_contents

tensor_ui

texture_size ğŸ”’

try_show_zoomed_image_region ğŸ”’

meter: iff this is a depth map, how long is one meter?

write_binary ğŸ”’