Crate re_data_ui

Rerun Data Ui

This crate provides ui elements for Rerun component data for the Rerun Viewer.

Modules

annotation_context ğŸ”’

app_id ğŸ”’

blob ğŸ”’

component ğŸ”’

component_name ğŸ”’

component_path ğŸ”’

component_ui_registry ğŸ”’

data_source ğŸ”’

entity_db ğŸ”’

entity_path ğŸ”’

image ğŸ”’

instance_path ğŸ”’

item_ui

Basic ui elements & interaction for most re_viewer_context::Item.

store_id ğŸ”’

tensor ğŸ”’

video ğŸ”’

Structs

ComponentPathLatestAtResults

All the values of a specific re_log_types::ComponentPath.

Traits

DataUi

Types implementing DataUi can display themselves in an [egui::Ui].

EntityDataUi

Similar to DataUi, but for data that is related to an entity (e.g. a component).

Functions

add_to_registry

Registers how to show a given component in the ui.

annotations

register_component_uis

sorted_component_list_for_ui

Sort components for display in the UI.

tensor_summary_ui_grid_contents