Trait re_data_ui::DataUi

pub trait DataUi {
    // Required method
    fn data_ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        verbosity: UiVerbosity,
        query: &LatestAtQuery
    );
}

Types implementing DataUi can display themselves in an [egui::Ui].

Required Methods

fn data_ui( &self, ctx: &mut ViewerContext<'_>, ui: &mut Ui, verbosity: UiVerbosity, query: &LatestAtQuery )

If you need to lookup something in the data store, use the given query to do so.