re_data_ui
Trait DataUi

pub trait DataUi {
    // Required method
    fn data_ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        ui_layout: UiLayout,
        query: &LatestAtQuery,
        db: &EntityDb,
    );

    // Provided method
    fn data_ui_recording(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        ui_layout: UiLayout,
    ) { ... }
}

Types implementing DataUi can display themselves in an [egui::Ui].

Required Methods

fn data_ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, ui_layout: UiLayout, query: &LatestAtQuery, db: &EntityDb, )

If you need to lookup something in the chunk store, use the given query to do so.

Provided Methods

fn data_ui_recording( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, ui_layout: UiLayout, )

Called Self::data_ui using the default query and recording.