Trait re_data_ui::DataUi

pub trait DataUi {
    // Required method
    fn data_ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        ui_layout: UiLayout,
        query: &LatestAtQuery,
        db: &EntityDb
    );
}

Types implementing DataUi can display themselves in an [egui::Ui].

Required Methods

fn data_ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, ui_layout: UiLayout, query: &LatestAtQuery, db: &EntityDb )

If you need to lookup something in the data store, use the given query to do so.