Trait re_data_ui::EntityDataUi

pub trait EntityDataUi {
    // Required method
    fn entity_data_ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        ui_layout: UiLayout,
        entity_path: &EntityPath,
        query: &LatestAtQuery,
        db: &EntityDb
    );
}

Similar to DataUi, but for data that is related to an entity (e.g. a component).

This is given the context of the entity it is part of so it can do queries.

Required Methods

fn entity_data_ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, ui_layout: UiLayout, entity_path: &EntityPath, query: &LatestAtQuery, db: &EntityDb )

If you need to lookup something in the data store, use the given query to do so.